IMPALA-12436: (Addendum) Fix BUILD_MULTI=true jenkins/build.sh
Quotes MULTI assignment.

Fixes a bootstrap race when the builder is first created by
bootstrapping during creation.

Switches from using rm --all-inactive (which incidentally needed --force
to avoid a prompt, and would fail to delete the impala-toolchain builder
if in use) to delete and recreate the builder when FORCE_REBUILD is set;
this is sufficient to clear the cache.

Skips deleting the builder at the end in case someone wants to rebuild
without FORCE_REBUILD.

Change-Id: Idb86f48c4e61cf271693f13cdf16fe1e99065393
Reviewed-on: http://gerrit.cloudera.org:8080/20497
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>