Add secondary S3 upload location for toolchain packages
All Apache Impala precommit activities happen in AWS EC2, in their
us-west-2 region, while the native-toolchain bucket hosting the
binary toolchain packages is located in us-west-1 for historical
reasons. This adds a constant cost to every single build launched on the
upstream infrastructure.

This patch adds a secondary upload location to the package upload logic
to allow a second, "mirroring" bucket to be configured, which is
expected to be closer to the build location network-wise. The mirror
bucket is currently expected to be located in EC2's us-west-2 region,
but this is not specified in the upload code.

The secondary upload location is expected in a new environment variable,
S3_MIRROR_BUCKET. If the variable is undefined or empty, the secondary
upload is silently skipped.

Note: One important feature of the canonical toolchain bucket,
s3://native-toolchain, is to store the source packages for the toolchain
components. This function is not duplicated in the mirror buckets, so
toolchain builds will still download their source packages from
s3://native-toolchain, regardless of any upload location (primary or
mirror) specified.

The patch fixes a random typo as well.

Tested by running a full toolchain build, uploading the binary packages
to two temporary test buckets in S3, then comparing the result across
the two buckets. When this passed, a core-mode test pass was executed on
the Impala upstream master branch.

Change-Id: I3aabaaf699dde865efe11e600f4e78aeca550378
Reviewed-on: http://gerrit.cloudera.org:8080/20549
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>