IMPALA-12445: Install the 'file' utility for Redhat9
The Redhat9 docker image currently doesn't have the
'file' utility installed, which means that the LLVM
build will not strip symbols from LLVM binaries.
This adds 'file' to the list of packages to install
for Redhat 9. It also adds 'file' to the list of
utilities that are expected to be installed.

Testing:
 - Built a Redhat 9 docker image and ran a toolchain build
 - Built all docker images and verified that the new
   check for the 'file' utility passes

Change-Id: If495ed3bdc5c15b796608e723502e44cca5a39ae
Reviewed-on: http://gerrit.cloudera.org:8080/20805
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>