IMPALA-12886: Bump GoogleTest version to 1.14.0
This is to unblock clang-tidy runs on ARM.
The current version of gtest seems to suffer from a heap corruption
problem that causes unpredictable segfaults in unified-be-tests.

Upgrading to GoogleTest 1.14.0 fixed this problem in test runs, likely
because previously existing global variables have been removed from
GoogleTest's code.

The patch removes the GOOGLETEST_VERSION specifier from the ORC library
build, as unit tests are not built for it.

Change-Id: Iaa52b4809a7c969e863518b182d2df63256f4c43
Reviewed-on: http://gerrit.cloudera.org:8080/21133
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>