IMPALA-12900: Build binutils with -O3
The binutils build happens before we have switched over to
using the toolchain compiler. This means that it also does
not set CFLAGS/CXXFLAGS. The default optimization level
for binutils is -O2. It is possible that we could get a bit
extra speed by using -O3, so this sets CFLAGS/CXXFLAGS to use
-O3 for binutils.

Testing:
 - Toolchain builds on x86_64 and ARM

Change-Id: I2e75db0759b4d3d4e6cc2ce929b1741808f1b771
Reviewed-on: http://gerrit.cloudera.org:8080/21145
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>