Add RE2 2024-03-01 and Abseil
This adds a build of the latest RE2 release 2024-03-01.
RE2 took a dependency on Abseil starting in 2023-06-01,
so this adds Abseil as well.

Abseil is a C++ library developed by Google that provides
various functionality including string libraries, hash
tables, etc. It is useful on its own.

Abseil's shared library build produces a large number of
shared libraries that depend on each other. This implements
a basic optimization of the symlink_required_libs function
to check the destination directory to see if the required
library has already been symlinked (or is otherwise present).
This makes a big difference in the time take for
symlink_required_libs for Abseil.

Testing:
 - Ran a toolchain build

Change-Id: Ie8e8a40cb0e5243b68a9abe35a106872bd9bf7f8
Reviewed-on: http://gerrit.cloudera.org:8080/21165
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>