IMPALA-12915: Build shared library of GTest
When building Impala with shared libraries, we currently link against
the static library of GTest in both libkudu_test_util.so and
unifiedbetests. libkudu_test_util.so is also linked by unifiedbetests.
When unifiedbetests exits, both binaries try to delete the global
variables of GTest, which leads to a double-free memory issue.

This patch builds both the shared and static library of GTest so Impala
can pick the corresponding library correctly.

Test:
 - Built GTest locally and verify the issue in Impala got resolved

Change-Id: Ie34023fe4fb0691fbca1dc547d1e40ada1beca22
Reviewed-on: http://gerrit.cloudera.org:8080/21161
Tested-by: Quanlong Huang <huangquanlong@gmail.com>
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>