IMPALA-13020: Use 64-bit integer for Thrift max message size on C++
Currently, Thrift's max message size is specified with a 32-bit
signed integer, so it maxes out at 2GB. Impala has use cases
that can produce messages larger than 2GB, so this patches Thrift
to change the max message size to be an int64_t. This will allow
Impala to specify a limit larger than 2GB. This only applies to
Thrift's C++ code and it does not change Java.

Change-Id: I94fdd7b07fcc8dca0639839b40a9eff815090835
Reviewed-on: http://gerrit.cloudera.org:8080/21343
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>