IMPALA-13072: Increase verbosity to print compilation commands
It is useful to be able to examine the compilation commands
to verify that custom flags are set (e.g. some components
use -fno-omit-frame-pointer, etc). Components that use
CMake often do not print the compilation command.

This modifies make invocations to pass in VERBOSE=1, which
prints the command being executed. This only passes in
VERBOSE=1 for make invocations that are doing compilation,
so it skips some locations that are only doing "make install"
after already compiling the project (or are header-only
projects).

Some projects don't respect VERBOSE=1. Boost has a special
-d+2 flag that turns on verbose output. Other projects like
CURL use configure's --disable-silent-rules flag to enable
verbose output.

The output is not too big and it gets redirected to a
file, so this does not make it configurable.

Testing:
 - Ran a toolchain build and verified that the compilation
   commands are in the output

Change-Id: I1a27bfe401c5def341b2e52f7d5e77d7f0e3a19e
Reviewed-on: http://gerrit.cloudera.org:8080/21470
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>