IMPALA-13072: Add retries for s3 uploads to combat flakiness
On ARM toolchain builds, we have seen some uploads to s3 fail
with a segementation fault. Given the number of artifacts that
the toolchain uploads, even a relatively low error rate can
make it hard to get a passing build. This modifies the s3
upload code to retry up to 10 times to avoid this flakiness.

Testing:
 - Ran an ARM toolchain build and saw the retry happen
   successfully
 - Ran a toolchain build with an invalid s3 bucket and verified
   it failed after 10 retries

Change-Id: I95d858c99e965730303c2bfd90478ac5f68acf83
Reviewed-on: http://gerrit.cloudera.org:8080/21421
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>