IMPALA-13121: Switch to ccache 3.7.12
The docker images currently build and use ccache 3.3.3.
Recently, we ran into a case where debuginfo was being
generated even though the cflags ended with -g0. The
ccache release history has this note for 3.3.5:
 - Fixed a regression where the original order of
   debug options could be lost.

This upgrades ccache to 3.7.12 to address this issue.

Ccache 3.7.12 is the last ccache release that builds
using autotools. Ccache 4 moves to build with CMake.
Adding a CMake dependency would be complicated at this
stage, because some of the older OSes don't provide a
new enough CMake in the package repositories. Since we
don't really need the new features of Ccache 4+, this
sticks with 3.7.12 for now.

This reenables the check_ccache_works() logic in
assert-dependencies-present.py.

Testing:
 - Built docker images and ran a toolchain build
 - The newer ccache resolves the unexpected debuginfo issue

Change-Id: I90d751445daa0dc298b634c1049d637a14afac40
Reviewed-on: http://gerrit.cloudera.org:8080/21473
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>