IMPALA-13309: Workaround Kudu build issue
A Maven repository shut down and Kudu's build fails trying
to find the gradle-scalafmt artifact. Maven Central has
this artifact at a different path, so this changes the Kudu
build to patch Kudu's gradle file to point to the working
gradle-scalafmt path.

Newer Kudu versions don't have this problem, but the change
that fixed this issue introduced a separate issue (see
KUDU-3610). So, this fixes the build until we can upgrade.

Testing:
 - Ran a toolchain build, built against the new toolchain

Change-Id: I155e00431d235a92567c1415aa38b07d1dc6f3ef
Reviewed-on: http://gerrit.cloudera.org:8080/21713
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>