IMPALA-13406 (fixup): Drop old version of CURL
The change to switch to Curl 8.10.1 added --with-gssapi
to the configure command for Curl. On SLES15 with Curl 7.78,
this causes configure to fail with a cryptic message like this:
configure: error: --with-openssl was given but OpenSSL could not be detected

Curl 8.10.1 does not have this issue, so this drops the older
version of Curl. We won't need it after switching to Curl 8.10.1.

Testing:
 - Build locally with SLES15
 - Ran build for all x86_64 distros

Change-Id: Ie29b314eb665fe57dff954bb3c41bc129dee486d
Reviewed-on: http://gerrit.cloudera.org:8080/21886
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>