IMPALA-14702 (prep): Bump Abseil version and only build static
Abseil had some issues in its CMake build that resulted in
flakiness building against it. This bumps Abseil to a more
recent version (20250512.2) with those CMake build fixes.
Moving to a newer Abseil breaks the 2024-03-01 re2 build.
Since we aren't using that re2 version currently, this bumps
that re2 build to 2025-11-05. (The regular re2 version that
doesn't depend on Abseil continues to stay the same.)

Google tcmalloc has a dependency on Abseil. Abseil has many
small libraries with dependencies on each other. Consuming
the CMake files in lib/cmake/* makes this much easier.
We currently do a static build and then a shared build, but
that means that the shared build overwrites the CMake files
in lib/cmake/* so that it always uses shared libraries. That
makes it harder to ship it in a Docker build. This removes
the shared library build so that Abseil is always a static
library.

Testing:
 - Built google tcmalloc and verified that it uses static
   libraries for Abseil

Change-Id: Ie8f4a89ba44994b06b24b8fdef2d054a965cbd31
Reviewed-on: http://gerrit.cloudera.org:8080/24015
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>