IMPALA-14129: Patch hadoop-client to disable repository.apache.org
This applies a patch on top of hadoop that disables two
Maven repositories: repository.jboss.org and repository.apache.org
The build does not actually need those repositories. All the
artifacts needed for this build are available from central.
Repeated requests to repository.apache.org are discouraged and
can lead to an IP being banned.

Applying a patch changes some of the directories names, so this
needed further adjustments to handle that.

Testing:
 - Ran ARM build and used that to build Impala on ARM
 - Verified that the hadoop-client build did not access
   repository.apache.org based on the logs

Change-Id: I2a441c1dc2c43e5fdcd467486b50e531daff62eb
Reviewed-on: http://gerrit.cloudera.org:8080/22992
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>