IMPALA-14139: Extend toolchain build to Ubuntu 24.04
The patch adds the complete toolchain build infrastructure to the Impala
toolchain project:
- defines the toolchain build Docker container
- adds Ubuntu 24.04 to various OS lists in the build scripts
- adds Ubuntu 24.04 to the README

It also adds a patch to GDB 12.1, fixing a function signature mismatch
between a source file and its matching header file specific to ARM CPUs.
The mismatch broke the GDB build on ARM for Ubuntu 24.04.
The patch is an unmodified backport of commit 8b1a24e546f from
https://sourceware.org/git/binutils-gdb.git, which fixes this exact
compilation problem.

Change-Id: I3224e6b19863deaca7dd8e3590daaf44f0f6779f
Reviewed-on: http://gerrit.cloudera.org:8080/23134
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>