IMPALA-13235: Update Opentelemetry C++ SDK Build Options
The BUILD_SHARED_LIBS option is now OFF as using the shared library
was forcing that library to be added to LD_LIBRARY_PATH.

The WITH_THREAD_INSTRUMENTATION_PREVIEW option is now ON as this
experimental option provides hooks for logging detailed information
about the internal threads of the OTLP HTTP exporter. Using this
feature requires adding specific code to activate it.

The WITH_OTLP_FILE option is now ON so the static libraries build
with support for exporting telemetry data to a file (which will be
used during Impala testing).

Change-Id: I46c53184204a0feb87b391ff7fe949bdb2812c58
Reviewed-on: http://gerrit.cloudera.org:8080/23024
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>