IMPALA-14394: Bump Kudu to their 1.18.x branch
Pick up the most recent release (candidate) from Kudu.

Although Kudu 1.18 is already released, this commit bumps the
toolchain's version of Kudu to a somewhat newer branch, where work is
still in progress.

This is a conscious choice: unfortunately Kudu 1.18 was published with a
Java dependency problem that breaks Impala builds. This problem is fixed
in the currently added state of Kudu's branch-1.18.x, but 1.18.1 is not
published yet, as some OpenSSL-related patches are still being
considered for the 1.18.1 version.

This version already enables a stable Impala build; further Kudu updates
will be imported to the toolchain when the final Kudu 1.18.1 version
becomes available.

This also adds the Perl FindBin package to the redhat9 image, required
by the Kudu build process. Red Hat 9 packages Perl v5.32 (or newer),
which split out this package from the main Perl package, so now it
needs to be added separately.

Change-Id: Icd7b4aed7ed0661d1ac2eaef763c7447f1f09797
Reviewed-on: http://gerrit.cloudera.org:8080/23131
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>