IMPALA-14393: Bump binutils to 2.42, add its dependencies
Bump binutils to 2.42, which is the version shipping on Ubuntu 24.04.
This avoids linker warnings when linking to system libraries on other
recent platforms, (e.g. Ubuntu 22.04 or RHEL 9), where system libraries
contain a new property type 0xc0000002 in the .note.gnu.property
section.

Add the texinfo package for Red Hat compatible platforms (Ubuntu images
already install this image). The presence of this package is required
by the new binutils version: absence of the package breaks the binutils
build, although the package will not be used on most platforms, even if
it is present.

Texinfo pulls in the Red Hat PowerTools repo as a dependency, so this
also patches the redhat8.df Dockerfile to point the PowerTools repo
to the already used mirror site.

Add '-std=gnu++11' to CXXFLAGS for the binutils build on CentOS 7:
binutils starting with version 2.42 requires a C++11 compatible compiler
for a correct build, because it added char_16t and char32_t types to
the 'gold' linker. The system GCC version for CentOS / Red Hat 7 is
v4.8.5, which runs in C++89 mode by default, so it needs this command-
line flag to set the language level. The added option is benign for all
other platforms.

Change-Id: I059a83d9fcbe48ecd48a720951e75b028d879faa
Reviewed-on: http://gerrit.cloudera.org:8080/23130
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Laszlo Gaal <laszlo.gaal@cloudera.com>