IMPALA-14449: Pin package repos for redhat9 build images to 9.2
Package repos used during the toolchain build container creation for the
redhat9 platform were not pinned to the specific version of the base
image (Rocky 9.2). This caused newer versions of several packages to be
pulled in (or upgraded) during the package installation step of the
image generation. This was especially problematic for OpenSSL, which
was rebased from 3.0.x to 3.2.2 between RHEL and Rocky Linux releases
9.4 and 9.5.

Installing the newer release into the toolchain build image caused an
unintended hard dependency for Impala on OpenSSL 3.2, which prevented
Impala builds from completing on earlier minor versions of Rocky 9.

This patch pins the package repos for the redhat9 container build,
pinning the OS package versions, including that of OpenSSL implicitly.

Change-Id: I30862642b2498ffad2c8c0d0913c6797b7d22e71
Reviewed-on: http://gerrit.cloudera.org:8080/23442
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>