IMPALA-14760: Add Clang 12 for compilation
Impala's dependency on LLVM is very performance sensitive,
but we also use Clang for sanitizer builds. In that context,
our old version of Clang has been causing issues compiling
new dependencies (e.g. IMPALA-13190 for roaring bitmap,
IMPALA-14092 for Paimon's use of Arrow). This adds a build
of LLVM 12 to provide a newer Clang for our sanitizer
builds. This includes LLD in the build, which may be useful
as an alternative linker.

This does a lot of changes to the structure of the LLVM
build scripts. It adds an ability to build with PGO (which
is not used for this change). It also changes the LLVM
build to use ninja to limit link concurrency.

Change-Id: Iabd1ac45c4c3e28bc35288129b74fa754956a266
Reviewed-on: http://gerrit.cloudera.org:8080/24499
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>