IMPALA-14609: Add ARROW_VERSION=15.0.0 to support aarch64
The patch fix the jni loading issue for aarch64 environment.
impalad fails to start after building in aarch64 environments
with error:
```
E   IllegalStateException: error loading native libraries:
java.io.FileNotFoundException: aarch_64/libarrow_cdata_jni.so
```
this is because in arrow 13.0.0,arrow jni so is not available
in maven central repo. It is available in arrow 14.0.0 or later.
so we propose to add ARROW_VERSION=15.0.0 to fix this
jni loading issue.

Testing:
Built new toolchain and pass paimon test in aarch64 environment.

Change-Id: Ie040c0ef171d281b1cc1ad929e61cc817a9e4210
Reviewed-on: http://gerrit.cloudera.org:8080/23754
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>
Tested-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>